requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
ccxt==4.1.40
pandas==2.1.3
numpy==1.25.2
discord-webhook==1.3.0
discord.py==2.3.2
cryptography==41.0.7
pyngrok==7.0.0
gunicorn==21.2.0
flask-cors==4.0.0
aiohttp==3.8.6
websockets==11.0.3
schedule==1.2.0
PyJWT==2.8.0
colorama==0.4.6
